% Auto-generated: do not edit by hand
\name{''TouchableComponent}

\alias{''TouchableComponent}

\title{TouchableComponent component}

\description{
ExampleComponent is an example component. It takes a property, `label`, and displays it. It renders an input with the property `value` which is editable by the user.
}

\usage{
''TouchableComponent(children=NULL, id=NULL, className=NULL, n_clicks=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. }

\item{id}{Character. }

\item{className}{Character. }

\item{n_clicks}{Numeric. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

